requirements.txt
nltk==3.8.1
numpy==1.24.1
pandas==2.2.2
prettytable==3.10.0
pytorch_crf==0.7.2
scikit_learn==1.3.1
seqeval==1.2.2
six==1.16.0
torch==2.0.1+cu118
tqdm==4.66.1
transformers==4.34.0
